# Challenge-02/ms-orders/src/main/resources/static.documents/openapi.yaml
openapi: 3.0.3
info:
  title: Product service api
  description: |
   - [Product api allows the user to manipulate product data, searching by id, returning all registered products, modifying, deleting or creating new products.]
  version: 1.0
servers:
  - url: "http://localhost:8080"
    description: Generated server url
paths:
 /products:
  get:
      summary: List of products
      operationId: getAllProducts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
              example:
                - id: 1
                  name:The Witcher 3: Wild Hunt
                  value: 29.98
                  description: Embarque em uma épica jornada de caça a monstros. Escolhas impactam o destino neste RPG imersivo.
                - id: 2
                  name: Overwatch 2
                  value: 0.0
                  description: Batalhas frenéticas entre heróis em equipes. Estratégia e habilidade se unem neste shooter colorido.
        '404':
          description: not found


  post:
      summary: Add a new product to the database
      operationId: creatProduct
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the product
                value:
                  type: number
                  format: double
                  description: Value of the product
                description:
                  type: string
                  description: Description of the products
            example:
                id: 10
                name: The Last Of Us Part 1
                value: 320.00
                description: The game takes place twenty years after the outbreak of infection caused by a mutation of the Cordyceps fungus that destroyed much of human civilization.

        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Bad Request
        '409':
          description: Conflict


 /products/{Id}:
  get:
     summary: Find product by ID
     description: Returns a single product
     operationId: getProductById
     parameters:
       - name: Product Id
         in: path
         description: ID of product to return
         required: true
         schema:
            type: integer
            format: int64
     responses:
         '200':
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProductDTO'
                example:
                  id: 1
                  name: "The Witcher 3: Wild Hunt"
                  value: 29.98
                  description: Embarque em uma épica jornada de caça a monstros. Escolhas impactam o destino neste RPG imersivo.


         '400':
            description: Bad Request
         '404':
           description: not found


  put:
      summary: Updates a product
      operationId: updateProduct
      requestBody:
         content:
           application/json:
             schema:
               type: object
               properties:
                 name:
                   type: string
                   description: Name of the product
                 value:
                   type: number
                   format: double
                   description: Value of the product
                 description:
                   type: string
                   description: Description of the products

      parameters:
        - name: product Id
          in: path
          description: ID of the product that needs to be updated, passed in the url
          required: true
          schema:
            type: integer
            format: int64


      responses:
        '200':
          description: OK

        '404':
          description: not found

        '409':
          description: conflict


  delete:
      summary: Deletes a product
      operationId: deleteProduct
      parameters:

        - name: productId
          in: path
          description: Product id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK

        '400':
          description: Bad Request

components:
  schemas:
    ProductDTO:
    required:
      - name,
      - value,
      - description
    type: object
    properties:
      name:
        type: string
      value:
        type: number
      description:
        type: string